Add support for specifying targets using helpful names
Previously users had to specify the full GCC and kernel targets,
which are not always clear. By adding helpful target names such
as rpi1, users can target a given piece of hardware without
needing to know all the options.

The build script has also been tweaked so that it saves the source
tarballs outside the build tree, which means that they only need
to be downloaded once, and the same tarballs can be used for
multiple cross compilers. This saves a significant amount of time
on slower connections, as there are over 200MB of source tarballs
to download.